Use in-line literals for `LocalValue` conversions, use a separate fixture
