+++ dotnet/test/common/BiDi/Script/CallFunctionLocalValueTest.cs
@@ -378,81 +378,4 @@ class CallFunctionLocalValueTest : BiDiTestFixture
             """, false, new() { Arguments = [arg] });
         }, Throws.Nothing);
     }
-
-    [Test]
-    public void CanConvertNullBoolToLocalValue()
-    {
-        bool? arg = null;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<NullLocalValue>());
-    }
-
-    [Test]
-    public void CanConvertTrueToLocalValue()
-    {
-        bool arg = true;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<BooleanLocalValue>());
-        Assert.That((result as BooleanLocalValue).Value, Is.True);
-    }
-
-    [Test]
-    public void CanConvertFalseToLocalValue()
-    {
-        bool arg = false;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<BooleanLocalValue>());
-        Assert.That((result as BooleanLocalValue).Value, Is.False);
-    }
-
-    [Test]
-    public void CanConvertNullIntToLocalValue()
-    {
-        int? arg = null;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<NullLocalValue>());
-    }
-
-    [Test]
-    public void CanConvertZeroIntToLocalValue()
-    {
-        int arg = 0;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<NumberLocalValue>());
-        Assert.That((result as NumberLocalValue).Value, Is.Zero);
-    }
-
-    [Test]
-    public void CanConvertNullDoubleToLocalValue()
-    {
-        double? arg = null;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<NullLocalValue>());
-    }
-
-    [Test]
-    public void CanConvertZeroDoubleToLocalValue()
-    {
-        double arg = 0;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<NumberLocalValue>());
-        Assert.That((result as NumberLocalValue).Value, Is.Zero);
-    }
-
-    [Test]
-    public void CanConvertNullStringToLocalValue()
-    {
-        string arg = null;
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<NullLocalValue>());
-    }
-
-    [Test]
-    public void CanConvertStringToLocalValue()
-    {
-        string arg = "value";
-        var result = (LocalValue)arg;
-        Assert.That(result, Is.TypeOf<StringLocalValue>());
-        Assert.That((result as StringLocalValue).Value, Is.EqualTo(arg));
-    }
 }
